Final Commit

+++ src/main/java/com/fujitsu/dto/UserDto.java
@@ -1,13 +1,14 @@
 package com.fujitsu.dto;
 
 import com.fujitsu.model.Gender;
+import lombok.Builder;
 import lombok.Data;
 
 import java.util.Date;
 
 @Data
 public class UserDto {
-    private int id;
+    private Long id;
     private String nickName;
     private String password;
     private String emailAddress;
